refactor: implement modal in MoodPanel and standardize mood emojis
- Refactored MoodPanel to use modal pattern consistent with other panels
- Extracted MoodForm into standalone component with forwardRef pattern
- Created getMoodInfo utility to standardize mood emoji rendering across app
- Fixed mood buttons grid layout to properly display all emojis
- Updated DailySummary to use the shared emoji mapping
- Enhanced mood.css with improved styling for emoji grid and interactions
- Create the alias '@utils' in vite.config.js for easy import of utility functions
